requirements.txt
absl-py==2.1.0
aiohttp==3.9.3
aiosignal==1.3.1
alabaster==0.7.16
alembic==1.13.1
annotated-types==0.6.0
app-model==0.2.4
appdirs==1.4.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
Babel==2.14.0
build==1.0.3
cachetools==5.3.3
cachey==0.2.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorlog==6.8.2
comm==0.2.1
contourpy==1.2.0
cuda-python==12.5.0
cudf-cu12==24.6.0
cuml-cu12==24.6.1
cupy-cuda12x==13.2.0
cycler==0.12.1
Cython==3.0.10
dask==2024.5.1
dask-cuda==24.6.0
dask-cudf-cu12==24.6.0
dask-expr==1.1.1
debugpy==1.8.1
decorator==4.4.2
distributed==2024.5.1
distributed-ucxx-cu12==0.38.0
docstring-parser==0.15
docutils==0.20.1
exceptiongroup==1.2.0
executing==2.0.1
fastrlock==0.8.2
filelock==3.13.1
fonttools==4.49.0
freetype-py==2.4.0
frozenlist==1.4.1
fsspec==2024.2.0
greenlet==3.0.3
grpcio==1.65.0
HeapDict==1.0.1
hsluv==5.0.4
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.5.1
imagesize==1.4.1
imgaug==0.4.0
importlib-metadata==7.0.1
in-n-out==0.1.9
ipykernel==6.29.2
ipython==8.21.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
lazy_loader==0.3
libucx-cu12==1.15.0.post1
lightning-utilities==0.11.1
llvmlite==0.42.0
locket==1.0.0
magicgui==0.8.1
Mako==1.3.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
mypy-extensions==1.0.0
napari==0.4.19
napari-console==0.0.9
napari-plugin-engine==0.2.0
napari-plugin-manager==0.1.0a2
napari-svg==0.1.10
nest-asyncio==1.6.0
networkx==3.2.1
npe2==0.7.4
numba==0.59.0
numpy==1.26.4
numpydoc==1.6.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
nvtx==0.2.10
opencv-python==4.9.0.80
optuna==3.6.0
packaging==23.2
pandas==2.2.0
parso==0.8.3
partd==1.4.1
pexpect==4.9.0
pillow==10.2.0
Pint==0.23
platformdirs==4.2.0
pooch==1.8.0
proglog==0.1.10
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
psygnal==0.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pyconify==0.1.6
pydantic==2.6.1
pydantic-compat==0.1.2
pydantic_core==2.16.2
Pygments==2.17.2
pylibraft-cu12==24.6.0
pynvjitlink-cu12==0.2.4
pynvml==11.4.1
PyOpenGL==3.1.7
pyparsing==3.1.1
pyproject_hooks==1.0.0
PyQt5==5.15.9
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyqt5-tools==5.15.9.3.3
python-dateutil==2.8.2
python-dotenv==1.0.1
pytorch-lightning==2.2.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
qtconsole==5.5.1
QtPy==2.4.1
raft-dask-cu12==24.6.0
rapids-dask-dependency==24.6.0
referencing==0.33.0
requests==2.31.0
rich==13.7.0
rmm-cu12==24.6.0
rpds-py==0.17.1
scikit-image==0.22.0
scipy==1.12.0
shapely==2.0.3
six==1.16.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
Sphinx==7.2.6
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
SQLAlchemy==2.0.29
stack-data==0.6.3
superqt==0.6.1
sympy==1.12
tabulate==0.9.0
tblib==3.0.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorrt==8.6.1.post1
tensorrt-bindings==8.6.1
tensorrt-libs==8.6.1
tifffile==2024.2.12
tomli==2.0.1
tomli_w==1.0.0
toolz==0.12.1
torch==2.2.0
torch_tensorrt==2.2.0
torchaudio==2.2.0
torchmetrics==1.3.2
torchvision==0.17.0
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
treelite==4.1.2
triangle==20230923
triton==2.2.0
typer==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
ucx-py-cu12==0.38.0
ucxx-cu12==0.38.0
urllib3==2.2.0
video2images==1.3
vispy==0.14.1
wcwidth==0.2.13
Werkzeug==3.0.3
wrapt==1.16.0
yarl==1.9.4
zict==3.0.0
zipp==3.17.0
